casts and double frees

+++ Kernel/idt/syscalls.c
@@ -133,7 +133,9 @@ uint64_t sys_create_process(uint64_t code_ptr, uint64_t args_ptr, uint64_t name_
 uint64_t sys_kill_process(uint64_t pid, uint64_t retval, uint64_t _unused1, uint64_t _unused2, uint64_t _unused3, uint64_t _unused4)
 {
     int32_t result = kill_process((uint16_t)pid, (int32_t)retval);
-    return (uint64_t)result;
+    // Sign-extend int32_t to int64_t first, then cast to uint64_t
+    // This ensures -1 becomes 0xFFFFFFFFFFFFFFFF, not a random value
+    return (uint64_t)(int64_t)result;
 }
 
 uint64_t sys_get_pid(uint64_t _unused1, uint64_t _unused2, uint64_t _unused3, uint64_t _unused4, uint64_t _unused5, uint64_t _unused6)
@@ -151,22 +153,22 @@ uint64_t sys_set_priority(uint64_t pid, uint64_t new_priority, uint64_t _unused1
 {
     // Validate priority
     if (new_priority >= NUM_PRIORITIES)
-        return -1;
+        return (uint64_t)(int64_t)-1;
 
     int8_t result = set_priority((uint16_t)pid, (uint8_t)new_priority);
-    return (uint64_t)result;
+    return (uint64_t)(int64_t)result;
 }
 
 uint64_t sys_block(uint64_t pid, uint64_t _unused1, uint64_t _unused2, uint64_t _unused3, uint64_t _unused4, uint64_t _unused5)
 {
     int8_t result = set_status((uint16_t)pid, BLOCKED);
-    return (uint64_t)result;
+    return (uint64_t)(int64_t)result;
 }
 
 uint64_t sys_unblock(uint64_t pid, uint64_t _unused1, uint64_t _unused2, uint64_t _unused3, uint64_t _unused4, uint64_t _unused5)
 {
     int8_t result = set_status((uint16_t)pid, READY);
-    return (uint64_t)result;
+    return (uint64_t)(int64_t)result;
 }
 
 // Memory management syscalls
@@ -216,48 +218,48 @@ uint64_t sys_sem_init(uint64_t sem_id, uint64_t initial_value, uint64_t _unused1
 {
     sem_t id = (sem_t)sem_id;
     int8_t result = sem_init(&id, (uint32_t)initial_value);
-    return (uint64_t)result;
+    return (uint64_t)(int64_t)result;
 }
 
 uint64_t sys_sem_open(uint64_t sem_id, uint64_t _unused1, uint64_t _unused2, uint64_t _unused3, uint64_t _unused4, uint64_t _unused5)
 {
     sem_t id = (sem_t)sem_id;
     int8_t result = sem_open(&id);
-    return (uint64_t)result;
+    return (uint64_t)(int64_t)result;
 }
 
 uint64_t sys_sem_close(uint64_t sem_id, uint64_t _unused1, uint64_t _unused2, uint64_t _unused3, uint64_t _unused4, uint64_t _unused5)
 {
     sem_t id = (sem_t)sem_id;
     int8_t result = sem_close(&id);
-    return (uint64_t)result;
+    return (uint64_t)(int64_t)result;
 }
 
 uint64_t sys_sem_destroy(uint64_t sem_id, uint64_t _unused1, uint64_t _unused2, uint64_t _unused3, uint64_t _unused4, uint64_t _unused5)
 {
     sem_t id = (sem_t)sem_id;
     int8_t result = sem_destroy(&id);
-    return (uint64_t)result;
+    return (uint64_t)(int64_t)result;
 }
 
 uint64_t sys_sem_wait(uint64_t sem_id, uint64_t _unused1, uint64_t _unused2, uint64_t _unused3, uint64_t _unused4, uint64_t _unused5)
 {
     sem_t id = (sem_t)sem_id;
     int8_t result = sem_wait(&id);
-    return (uint64_t)result;
+    return (uint64_t)(int64_t)result;
 }
 
 uint64_t sys_sem_post(uint64_t sem_id, uint64_t _unused1, uint64_t _unused2, uint64_t _unused3, uint64_t _unused4, uint64_t _unused5)
 {
     sem_t id = (sem_t)sem_id;
     int8_t result = sem_post(&id);
-    return (uint64_t)result;
+    return (uint64_t)(int64_t)result;
 }
 
 uint64_t sys_waitpid(uint64_t pid, uint64_t _unused1, uint64_t _unused2, uint64_t _unused3, uint64_t _unused4, uint64_t _unused5)
 {
     int32_t result = waitpid((uint16_t)pid);
-    return (uint64_t)result;
+    return (uint64_t)(int64_t)result;
 }
 
 // Pipe syscalls
